refactored to match with the rest of the codebase
