--- src/ripplegw/waveforms/__init__.py
@@ -1,1 +1,0 @@
-from .sine_gaussian import SineGaussian
